@@ -2,9 +2,14 @@
   "name": "oslicense",
   "version": "1.2.1",
   "description": "A simple helper tool to retrieve Open Source Initiative-approved licenses.",
-  "main": "oslicense.js",
+  "main": "src/oslicense.js",
+  "type": "module",
+  "imports": {
+      "#app/*": "./src/*.js",
+      "#root/*": "./*.js"
+  },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=14.6.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
